fix: normalize qualified model labels
