+++ apps/mobile/src/lib/modelOptions.ts
@@ -6,6 +6,7 @@ import type {
 import {
   buildExplicitProviderOptionSelectionsFromDescriptors,
   getProviderOptionDescriptors,
+  stripLeadingModelQualifier,
 } from "@t3tools/shared/model";
 
 export type ModelOption = {
@@ -44,7 +45,8 @@ function modelDisplayLabel(model: {
   readonly subProvider?: string | undefined;
 }): string {
   const subProvider = model.subProvider?.trim();
-  return subProvider ? `${model.name} · ${subProvider}` : model.name;
+  const modelName = stripLeadingModelQualifier(model.name, subProvider);
+  return subProvider ? `${modelName} · ${subProvider}` : modelName;
 }
 
 function normalizeSelectionOptions(
