+++ apps/web/src/components/chat/ProviderModelPicker.tsx
@@ -210,7 +210,14 @@ export const ProviderModelPicker = memo(function ProviderModelPicker(props: {
             >
               <span className="min-w-0 truncate">{props.triggerLabel ?? triggerTitle}</span>
               {triggerSubProvider && props.triggerLabel === undefined ? (
-                <span className="shrink-0 text-muted-foreground">· {triggerSubProvider}</span>
+                <span
+                  className={cn(
+                    "text-muted-foreground",
+                    props.compact ? "min-w-0 truncate" : "shrink-0",
+                  )}
+                >
+                  · {triggerSubProvider}
+                </span>
               ) : null}
             </TooltipTrigger>
             <TooltipPopup side="top">{props.triggerLabel ?? triggerLabel}</TooltipPopup>
