+++ apps/web/src/components/chat/providerIconUtils.ts
@@ -1,4 +1,5 @@
 import { ProviderDriverKind } from "@t3tools/contracts";
+import { stripLeadingModelQualifier } from "@t3tools/shared/model";
 import {
   AntigravityIcon,
   ClaudeAI,
@@ -30,26 +31,12 @@ export type ModelEsque = {
   isUnavailable?: boolean | undefined;
 };
 
-function escapeRegExp(value: string): string {
-  return value.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
-}
-
-function stripLeadingQualifier(value: string, qualifier: string | null | undefined): string {
-  const trimmedQualifier = qualifier?.trim();
-  if (!trimmedQualifier) {
-    return value;
-  }
-
-  const pattern = new RegExp(`^${escapeRegExp(trimmedQualifier)}(?:\\s*[.:/-]\\s*|\\s+)`, "iu");
-  return value.replace(pattern, "").trim() || value;
-}
-
 export function getDisplayModelName(
   model: ModelEsque,
   options?: { preferShortName?: boolean },
 ): string {
   const name = options?.preferShortName && model.shortName ? model.shortName : model.name;
-  return stripLeadingQualifier(name, model.subProvider);
+  return stripLeadingModelQualifier(name, model.subProvider);
 }
 
 export function getTriggerDisplayModelName(model: ModelEsque): string {
